requirements.txt
streamlit
matplotlib
mysql-connector-python
networkx
nltk
plotly
pmdarima
pydalvis
sastrawi
textblob
vadersentiment
